requirements.txt
aiofiles==0.6.0
APScheduler==3.7.0
bcrypt==3.2.0
coverage==6.1.2
decorator==5.0.9
faker==9.8.2
factory-boy==3.2.1
fastapi==0.70.0
fastapi-mail==1.0.2
fastapi-jwt-auth==0.5.0
PyJWT==1.7.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
SQLAlchemy==1.4.11
SQLAlchemy-Utils==0.37.9
schedule==1.1.0
uvicorn==0.13.4
python-dotenv==0.19.0
PyMySQL==1.0.2
passlib~=1.7.4
pydantic~=1.8.2
requests==2.26.0
pytest==6.2.5
opensearch-py==1.0.0